multimind/gateway/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
click>=8.0.0
rich>=10.0.0

# Model-specific dependencies
openai>=1.0.0
anthropic>=0.5.0
groq>=0.3.0
huggingface-hub>=0.16.0
requests>=2.26.0

# Testing dependencies
pytest>=6.2.5
pytest-asyncio>=0.16.0
httpx>=0.23.0
pytest-cov>=2.12.0

# Development dependencies
black>=21.9b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910 